Retrieve build dependencies in a separate dockerfile build step to the gradle build.
Downloading build dependencies can take quite some time and transfers a
large amount of data. By retrieving these dependencies in a separate
dockerfile build step compare to the gradle build, we give docker a
chance to detect that there are no dependency changes and that a
cached docker build step can be used with the dependencies already
downloaded.